requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Training orchestration
lightning>=2.0.0

# Hugging Face
transformers>=4.35.0

# LoRA fine-tuning (required only for --lora flag)
peft>=0.6.0

# Data
roboflow>=1.1.0
python-dotenv>=1.0.0

# Image / annotation processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pycocotools>=2.0.7

# Augmentation
albumentations>=1.3.1

# Utility
tqdm>=4.66.0
